+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Introduced JS module `wc/file/util`, it provides following file operations `blobToFile`, `fixFileExtension`, `dataURItoBlob` and `getMimeToExtMap`.
 
 ### Bug Fixes
+* Fixed WFileWidget and WMultiFileWidget to accept filetypes of both MIME and extension type #1584
 
 ## Release 1.5.13